pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.1",
    "black>=25.1.0",
    "bs4>=0.0.2",
    "dill>=0.4.0",
    "fastmcp==2.3.0",
    "ipython>=9.2.0",
    "mcp>=1.9.2",
    "numpy>=2.2.6",
    "openai==1.66.3",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "tavily-python>=0.7.3",
]
